requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
appdirs==1.4.4
astunparse==1.6.3
atomicwrites==1.4.0
azure-functions==1.17.0
black==23.9.1
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
constantly==15.1.0
cssselect==1.1.0
daal4py==2023.1.1
distlib==0.3.7
et-xmlfile==1.1.0
eventkit==1.0.3
filelock==3.12.4
financetoolkit==1.2.1
flake8
flatbuffers==23.5.26
fonttools==4.25.0
frozendict==2.3.8
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
html5lib==1.1
ib-insync==0.9.86
identify==2.5.30
idna==3.4
isort
jsonpointer==2.1
keras==2.14.0
keyring
libclang==16.0.6
mkl-service==2.4.1
ml-dtypes==0.2.0
matplotlib
mpmath==1.2.1
multitasking==0.0.11
munkres==1.1.4
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
pandas==2.0.3
patsy==0.5.3
pep8==1.7.1
Pillow==9.4.0
platformdirs==3.10.0
plotly
pluggy
ply==3.11
pre-commit==3.5.0
protobuf==4.24.3
pyarrow==11.0.0
pyasn1-modules==0.2.8
pydantic==2.6.4
pydantic_core==2.16.3
PyDispatcher==2.0.5
pydocstyle
pyls-spyder==0.4.0
PyNaCl
pyodbc
pyOpenSSL
pyparsing
PyQt5==5.15.7
PyQtWebEngine==5.15.4
pytest-env==1.0.1
pytest-mock==3.11.1
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
pytz==2023.3
queuelib==1.5.0
requests==2.31.0
requests-mock==1.11.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
simfin==0.9.2
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
termcolor==2.3.0
tzdata==2023.3
types-requests
urllib3==2.0.4
virtualenv==20.24.5
yfinance==0.2.26
zstandard==0.19.0
lseg.data
refinitiv.data